% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_osrm_limits}
\alias{check_osrm_limits}
\title{Check OSRM's query limits and provide warnings.}
\usage{
check_osrm_limits(src, dst)
}
\description{
This function is created in order to prevent the user
from composing queries that go over the \code{default} limits of the
OSRM servers.
}
\seealso{
\code{\link{osrmtable}} for the main function
}
\keyword{internal}
