% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{common_swift}
\alias{common_swift}
\title{Nests count of common swifts in Utrecht.}
\format{
A data frame with 99 rows and 2 variables:
\describe{
\item{nest_count}{number of nests in one position}
\item{geometry}{the geographical position of the nests}
...
}
}
\source{
\url{https://ckan.dataplatform.nl/dataset/gierzwaluwinventarisatie-2014}
}
\usage{
common_swift
}
\description{
An sf dataset containing 99 nests count of common swifts in the
municipality of Utrecht (The Netherlands) and their geographical position.
As the dataset is formatted as a simple feature collection it also contains
attributes such as the geometry_type, dimension, bbox and CRS.
}
\keyword{datasets}
