% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_osrm_input_table}
\alias{create_osrm_input_table}
\title{Build OSRM query}
\usage{
create_osrm_input_table(loc, osrm_server, osrm_profile)
}
\description{
Build OSRM query
}
\seealso{
\code{\link[=osrm_table]{osrm_table()}} for the main function
}
\keyword{internal}
