% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich.R
\name{enrich}
\alias{enrich}
\alias{enrich_osm}
\title{Enrich OSM Data}
\usage{
enrich_osm(
  dataset,
  name = NULL,
  key = NULL,
  value = NULL,
  type = "points",
  distance = "spherical",
  kernel = "uniform",
  ...,
  .verbose = TRUE
)
}
\arguments{
\item{dataset}{target \code{sf} dataset to enrich with this package}

\item{name}{the column name of the feature to be added}

\item{key}{target OSM feature key to add, see \link{add_osm_feature}}

\item{value}{target value for OSM feature key to add, see \link{add_osm_feature}}

\item{type}{\code{character} the osm feature type or types to consider
(e.g., points, polygons), see details}

\item{distance}{\code{character} the distance metric used, see details}

\item{kernel}{\code{function} the kernel function used, see details}

\item{...}{\code{enriched_overpass_query} column or columns to add}

\item{.verbose}{\code{bool} whether to print info during enrichment}
}
\description{
Perform enriched query on osm and add as new column.

The enrichment call works in the following way: an \code{enriched_overpass_query}
(e.g. \code{waste_query}) is created and then a new column is added by specifying
the name of the column (\code{"waste_baskets = waste_query"}). This call
also works with more than one queries.
}
\note{
If you want to get a large number of points make sure to set the
.timeout (time before request times out) and .memsize (maxmimum
size of the request) arguments for the Overpass server and set
the "max-table-size" argument correctly when starting the
OSRM server(s).
}
\examples{
\dontrun{

#' # Enrich data creating new column `waste_baskets`
sf_enriched <- dataset \%>\%
  enrich_osm(
    name = "waste_baskets",
    key = "amenity",
    value = "waste_basket",
    type = "points",
    distance = "walking_duration",
    kernel = "uniform",
    r = 100
  )
}

}
\seealso{
\code{\link{enrich_opq}}
}
