% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_osrm_output_table}
\alias{format_osrm_output_table}
\title{Format OSRM output tables}
\usage{
format_osrm_output_table(out, features, type)
}
\description{
Format OSRM output tables
}
\seealso{
\code{\link[=osrm_table]{osrm_table()}} for the main function
}
\keyword{internal}
