% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{kernels}
\alias{kernels}
\alias{kernel_gaussian}
\alias{kernel_parabola}
\alias{kernel_uniform}
\title{Kernels used in summarizing features}
\usage{
kernel_gaussian(d, r = 100, FUN = sum)

kernel_parabola(d, r = 100, FUN = sum)

kernel_uniform(d, r = 100, FUN = sum)
}
\arguments{
\item{d}{Vector of distances.}

\item{r}{Radius of interest.}

\item{FUN}{Reduce function.}
}
\value{
numeric, function (usually sum) of kernel-weighted distances
}
\description{
Kernels convert distance vectors to single numbers,
with a certain weight for certain distances. In general, at 0 distance,
these return a value of 1, and the value decreases as the distance
increases, reaching 0 at the limit of the support.
}
\details{
Gaussian kernel is a truncated gaussian, where r = 4*sigma
(i.e., std. dev = r/4).The density which is truncated away is
1 - erf(2 sqrt(2)), which is approximately 0.0000633.
Parabola kernel is parabolic, decreasing with distance within
radius r, and returns 0 elsewhere. It is a scaled Epanechnikov kernel.
Uniform kernel returns 1 within radius r, 0 elsewhere.
}
\seealso{
\link{enrich_opq}
}
