% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osrmtable.R
\name{osrmtable}
\alias{osrmtable}
\alias{osrm_table}
\title{Get Travel Time Matrices Between Points}
\usage{
osrm_table(
  src = NULL,
  dst = NULL,
  measure = "duration",
  osrm_profile = "driving",
  osrm_server = getOption("osrm.server")
)
}
\arguments{
\item{src}{a data frame containing origin points identifiers, longitudes
and latitudes (WGS84). It can also be a SpatialPointsDataFrame, a
SpatialPolygonsDataFrame or an sf object. If so, row names are used as identifiers.
If dst and src parameters are used, only pairs between scr/dst are computed.}

\item{dst}{a data frame containing destination points identifiers, longitudes
and latitudes (WGS84). It can also be a SpatialPointsDataFrame a
SpatialPolygonsDataFrame or an sf object. If so, row names are used as identifiers.}

\item{measure}{a character indicating what measures are calculated. It can
be "duration" (in minutes), "distance" (meters), or both c('duration',
'distance'). The demo server only allows "duration".}

\item{osrm_profile}{the routing profile to use, e.g. "car", "bike" or "foot"
(when using the routing.openstreetmap.de test server).
getOption("osrm.profile") by default.}

\item{osrm_server}{the base URL of the routing server.
getOption("osrm.server") by default.}
}
\value{
A list containing 3 data frames is returned.
durations is the matrix of travel times (in minutes),
sources and destinations are the coordinates of
the origin and destination points actually used to compute the travel
times (WGS84).
}
\description{
Build and send OSRM API queries to get travel time matrices
between points. This function interfaces the \emph{table} OSRM service.
}
\seealso{
\code{\link{set_server_profile}} and \code{\link{utils}} for
helper functions used in this function
}
