% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_server_profile}
\alias{set_server_profile}
\title{Test if server set is available.}
\usage{
set_server_profile(server, profile)
}
\arguments{
\item{profile}{The name of the profile}
}
\value{
Set the \code{osrm.server} and the \code{osrm.profile}
The name of the profile will be used only if the server (via \code{osrm.server})
is set. Otherwise, the default server will be used
}
\description{
The user should indicate the address where one or more instance
of the OSRM server are running. In case more instances are running at the
same time, we force the user to place them behind a reverse proxy (we force
to indicate only one server address). We further recommend to use our
default docker-compose configuration to have a working configuration
with three OSRM servers behind an nginx server
(https://github.com/sodascience/osmenrich_docker).

If no servers are used, the default server from the OSRM project
(http://router.project-osrm.org/) will be used. This is not recommended
as this server is intended only for demo purposes, might fail in case of
overload and only returns distances for the driving profile.
}
\seealso{
\code{\link{osrmtable}} for the main function
}
\keyword{internal}
