% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{transform_to_df}
\alias{transform_to_df}
\title{Take sf and transform it into a dataframe}
\usage{
transform_to_df(sf)
}
\description{
Dataframes are necessary for the transformations
needed to query the osrm servers.
}
\seealso{
\code{\link{osrmtable}}
}
\keyword{internal}
