% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{url_available}
\alias{url_available}
\title{Check if server is available}
\usage{
url_available(u)
}
\description{
Check if server is available
}
\seealso{
\code{\link[=set_server_profile]{set_server_profile()}}
}
\keyword{internal}
