% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{adjust_coord}
\alias{adjust_coord}
\title{Adjust coordinates to fit within OSRM requirements}
\usage{
adjust_coord(coord)
}
\description{
Adjust coordinates to fit within OSRM requirements
}
\seealso{
\code{\link{osrmtable}}
}
\keyword{internal}
