% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_osm.R
\name{enrich}
\alias{enrich}
\alias{enrich_osm}
\alias{data_enrichment}
\alias{distance_matrix}
\alias{control_enrich}
\title{Enrich \code{sf} object with OSM data}
\usage{
enrich_osm(
  dataset,
  name = NULL,
  key = NULL,
  value = NULL,
  type = "points",
  distance = "spherical",
  r = NULL,
  kernel = "uniform",
  reduce_fun = sum,
  control = list(),
  .verbose = TRUE,
  ...
)

data_enrichment(data, query, colname, .verbose = TRUE)

distance_matrix(distancename, distancefun, ref_geometry, ftr_geometry)

control_enrich(timeout = 300, memsize = 1073741824)
}
\arguments{
\item{dataset}{target \code{sf} dataset to enrich with this package}

\item{name}{the column name of the feature to be added}

\item{key}{target OSM feature key to add, see \code{\link[osmdata:add_osm_feature]{osmdata::add_osm_feature()}}}

\item{value}{target value for OSM feature key to add, see
\code{\link[osmdata:add_osm_feature]{osmdata::add_osm_feature()}}}

\item{type}{\code{character} the osm feature type or types to consider
(e.g., points, polygons), see details}

\item{distance}{\code{character} the distance metric used, see details}

\item{r}{The search radius used by the \code{kernel} function.}

\item{kernel}{\code{function} the kernel function used, see details}

\item{reduce_fun}{The aggregation function used by the \code{kernel} function to
aggregate the retrieved data points.}

\item{control}{The list with configuration variables for the OSRM server.
It contains \code{timeout}, defining the number of seconds before the request
to OSRM times out, and \code{memsize}, defining the maximum size of the query to
OSRM.}

\item{.verbose}{\code{bool} whether to print info during enrichment}

\item{...}{Additional parameters to be passed into the OSM query, such as
a user-defined kernel.}
}
\description{
Perform enriched query on OSM and add as new column.
}
\details{
\code{Type} represents the feature type to be considered. Usually this
would be points, but polygons and multipolygons are also possible. This
argument can also be a vector of multiple types. Non-point types will be
converted to points using the \code{st_centroid} function from the \code{sf} package
(NB this does not necessarily work well for all features!).
Available options are:
\itemize{
\item points
\item lines
\item polygons
\item multilines
\item multipolygons
}

\code{Distance} represents the metric used to compute the distances between the
rows in the dataset and the OSM features. \code{Duration} represents the metric
that indicates the average duration to cover the distances between the
rows in the dataset and the OSM features. The following metrics are
available in this package, assuming that the OSRM server is setup as
suggested in our guide at:
https://github.com/sodascience/osmenrich_docker:
\itemize{
\item spherical
\item distance_by_foot
\item duration_by_foot
\item distance_by_car
\item duration_by_car
\item distance_by_bike
\item duration_by_bike
}

\code{Kernel} is a kernel function from the \code{osmenrich} package to be used in
weighing the features and the radius/distance where features are
considered. For simply counting the number of occurrences within a radius,
use \code{kernel_uniform} with radius \code{r}.

For more details see the introductory vignette of \code{osmenrich}:
\code{vignette("introduction", package = "osmenrich")}
}
\note{
If you want to get a large number of points make sure to set the
\code{.timeout} (time before request times out) and \code{.memsize} (maxmimum
size of the request) arguments for the Overpass server and set
the "max-table-size" argument correctly when starting the
OSRM server(s).
}
\examples{
\donttest{
# Load libraries
library(tidyverse)
library(sf)

# Create example dataset
sf_example <-
tribble(
  ~person,  ~lat,   ~lon,
  "Alice",  52.12,  5.09,
  "Bob",    52.13,  5.08,
  ) \%>\%
  sf::st_as_sf(
  coords = c("lon", "lat"),
  crs = 4326
 )

# Enrich data creating new column `waste_baskets`
sf_enriched <- dataset \%>\%
  enrich_osm(
    name = "n_waste_baskets",
    key = "amenity",
    value = "waste_basket",
    reduce_fun = sum
  )
}

}
\seealso{
\code{\link[=enrich_opq]{enrich_opq()}}
}
\keyword{internal}
