% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opqenrich.R
\name{enrich_opq}
\alias{enrich_opq}
\alias{add_bbox}
\alias{add_feature}
\alias{add_type}
\alias{add_distance}
\alias{add_kernel}
\title{Enrich an overpass query for column output}
\usage{
enrich_opq(
  dataset,
  name = NULL,
  key = NULL,
  value = NULL,
  type = "points",
  distance = "spherical",
  kernel = "uniform",
  .verbose = TRUE,
  ...
)

add_bbox(dataset, ...)

add_feature(opq, key, value)

add_type(opq, type)

add_distance(opq, distance)

add_kernel(opq, kernel, ...)
}
\arguments{
\item{dataset}{Target \code{sf} dataset to enrich with this package.}

\item{name}{Name of the enriched column.}

\item{key}{Target OSM feature key to add, see \link{add_osm_feature}.}

\item{value}{Target value for OSM feature key to add, see \link{add_osm_feature}.}

\item{type}{\code{character} The OSM feature type or types to consider
(e.g., points, polygons), see details}

\item{distance}{\code{character} The distance metric used, see details.}

\item{kernel}{\code{function} The kernel function used, see details.}

\item{.verbose}{\code{bool} Whether to print info during enrichment.}

\item{...}{Arguments passed to the \code{kernel} function (\code{r} and \code{sd}).}

\item{opq}{Overpass query that is being enriched.}
}
\description{
Enrich an overpass query for column output
}
\seealso{
\link{enrich_osm}
}
