% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_opq.R
\name{enrich_opq}
\alias{enrich_opq}
\alias{add_bbox}
\alias{add_feature}
\alias{add_type}
\alias{add_distance}
\alias{add_kernel}
\title{Enrich an overpass query for column output}
\usage{
enrich_opq(
  dataset,
  name = NULL,
  key = NULL,
  value = NULL,
  type = "points",
  distance = "spherical",
  r = NULL,
  kernel = "uniform",
  reduce_fun = sum,
  control = list(),
  .verbose = TRUE,
  ...
)

add_bbox(dataset, r, control)

add_feature(opq, key, value)

add_type(opq, type)

add_distance(opq, distance)

add_kernel(opq, kernel, r, reduce_fun, ...)
}
\arguments{
\item{dataset}{target \code{sf} dataset to enrich with this package}

\item{name}{name of the enriched column}

\item{key}{target OSM feature key to add, see \code{\link[osmdata:add_osm_feature]{osmdata::add_osm_feature()}}}

\item{value}{target value for OSM feature key to add, see
\code{\link[osmdata:add_osm_feature]{osmdata::add_osm_feature()}}}

\item{type}{\code{character} the osm feature type or types to consider
(e.g., points, polygons), see details}

\item{distance}{\code{character} the distance metric used, see details}

\item{r}{The search radius used by the \code{kernel} function.}

\item{kernel}{\code{function} the kernel function used, see details}

\item{reduce_fun}{The aggregation function used by the \code{kernel} function to
aggregate the retrieved data points.}

\item{control}{The list with configuration variables for the OSRM server.
It contains \code{timeout}, defining the number of seconds before the request
to OSRM times out, and \code{memsize}, defining the maximum size of the query to
OSRM.}

\item{.verbose}{\code{bool} whether to print info during enrichment}

\item{...}{Additional parameters to be passed into the OSM query, such as
a user-defined kernel.}

\item{opq}{overpass query that is being enriched}
}
\description{
Enrich an overpass query for column output
}
\seealso{
\code{\link[=enrich_osm]{enrich_osm()}}
}
