% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_coord_table}
\alias{format_coord_table}
\title{Format output coordinates table}
\usage{
format_coord_table(out, features, type)
}
\description{
Format output coordinates table
}
\seealso{
\code{\link{osrmtable}} for the main function
}
\keyword{internal}
