% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osrmtable.R
\name{osrmtable}
\alias{osrmtable}
\alias{osrm_table}
\title{Get Travel Time Matrices Between Points}
\usage{
osrm_table(
  src = NULL,
  dst = NULL,
  measure = "duration",
  osrm_profile = "driving",
  osrm_server = getOption("osrm.server")
)
}
\arguments{
\item{src}{A data frame containing origin points identifiers, longitudes
and latitudes (WGS84). It can also be a \code{SpatialPointsDataFrame}, a
\code{SpatialPolygonsDataFrame} or an \code{sf} object. If so, row names are
used as identifiers. If dst and src parameters are used,
only pairs between scr/dst are computed.}

\item{dst}{A data frame containing destination points identifiers, longitudes
and latitudes (WGS84). It can also be a \code{SpatialPointsDataFrame} a
\code{SpatialPolygonsDataFrame} or an \code{sf} object. If so, row names are used
as identifiers.}

\item{measure}{A character indicating what measures are calculated. It can
be "duration" (in minutes), "distance" (meters), or both c('duration',
'distance'). The public server only allows "duration".}

\item{osrm_profile}{The routing profile to use, e.g. "car", "bike" or "foot"
(when not using the public server). Uses \code{getOption("osrm.profile")} to
retrieve the default profile ("car").}

\item{osrm_server}{The base URL of the routing server. Uses
\code{getOption("osrm.server")} to retrieve the current server by default.}
}
\value{
A list containing 3 data frames is returned.
\code{durations} is the matrix of travel times (in minutes);
\code{sources} and \code{destinations} are the coordinates of
the origin and destination points actually used to compute the travel
times (WGS84).
}
\description{
Build and send OSRM API queries to get travel time matrices
between points. This function interfaces the \emph{table} OSRM service.
}
\seealso{
\code{\link{set_server_profile}} and \code{\link{utils}} for
helper functions used in this function
}
